Complete GET method

+++ inc/CGI.hpp
@@ -17,8 +17,8 @@ public:
 	void	setPrintFD(int fd);
 	void	addQuery(std::string query_string);
 
-	virtual void	readEvent(int read_size);
-	virtual void	writeEvent(int write_size);
+	virtual void	readEvent(long read_size);
+	virtual void	writeEvent(long write_size);
 	virtual void	timerEvent();
 
 private:
